+++ crates/lgn-graphics-renderer/src/gpu_renderer/mod.rs
@@ -1,6 +1,3 @@
-mod gpu_instance_manager;
-pub(crate) use gpu_instance_manager::*;
-
 mod render_batch;
 pub(crate) use render_batch::*;
