+++ crates/lgn-graphics-renderer/src/gpu_renderer/render_element.rs
@@ -1,8 +1,9 @@
 use lgn_graphics_api::CommandBuffer;
 
-use crate::resources::{MaterialId, MeshTopology, RenderMesh};
-
-use super::GpuInstanceId;
+use crate::{
+    features::GpuInstanceId,
+    resources::{MaterialId, MeshTopology, RenderMesh},
+};
 
 #[derive(Clone, Copy)]
 pub struct RenderElement {
